[dashboard] Remove DataSource.events. (#49380)
