+++ python/ray/dashboard/modules/state/state_head.py
@@ -190,14 +190,6 @@ class StateHead(dashboard_utils.DashboardHeadModule, RateLimitedModule):
         record_extra_usage_tag(TagKey.CORE_STATE_API_LIST_RUNTIME_ENVS, "1")
         return await handle_list_api(self._state_api.list_runtime_envs, req)
 
-    @routes.get("/api/v0/cluster_events")
-    @RateLimitedModule.enforce_max_concurrent_calls
-    async def list_cluster_events(
-        self, req: aiohttp.web.Request
-    ) -> aiohttp.web.Response:
-        record_extra_usage_tag(TagKey.CORE_STATE_API_LIST_CLUSTER_EVENTS, "1")
-        return await handle_list_api(self._state_api.list_cluster_events, req)
-
     @routes.get("/api/v0/logs")
     @RateLimitedModule.enforce_max_concurrent_calls
     async def list_logs(self, req: aiohttp.web.Request) -> aiohttp.web.Response:
