+++ python/ray/dashboard/state_api_utils.py
@@ -1,6 +1,6 @@
 import dataclasses
 from dataclasses import asdict, fields
-from typing import Callable, List, Tuple
+from typing import Awaitable, Callable, List, Tuple
 
 import aiohttp.web
 
@@ -10,6 +10,7 @@ from ray.util.state.common import (
     DEFAULT_RPC_TIMEOUT,
     RAY_MAX_LIMIT_FROM_API_SERVER,
     ListApiOptions,
+    ListApiResponse,
     PredicateType,
     StateSchema,
     SummaryApiOptions,
@@ -21,7 +22,7 @@ from ray.util.state.exception import DataSourceUnavailable
 from ray.util.state.util import convert_string_to_type
 
 
-def do_reply(success: bool, error_message: str, result: dict, **kwargs):
+def do_reply(success: bool, error_message: str, result: ListApiResponse, **kwargs):
     return rest_response(
         success=success,
         message=error_message,
@@ -32,7 +33,8 @@ def do_reply(success: bool, error_message: str, result: dict, **kwargs):
 
 
 async def handle_list_api(
-    list_api_fn: Callable[[ListApiOptions], dict], req: aiohttp.web.Request
+    list_api_fn: Callable[[ListApiOptions], Awaitable[ListApiResponse]],
+    req: aiohttp.web.Request,
 ):
     try:
         result = await list_api_fn(option=options_from_req(req))
