+++ python/ray/tests/test_state_api.py
@@ -5,15 +5,19 @@ import sys
 import signal
 from collections import Counter
 from concurrent.futures import ThreadPoolExecutor
-from typing import List, Tuple
+from typing import List
 from unittest.mock import MagicMock, AsyncMock, patch
 
 import pytest
 import pytest_asyncio
-from ray._private.state_api_test_utils import get_state_api_manager
+from ray._private.state_api_test_utils import (
+    get_state_api_manager,
+    create_api_options,
+    verify_schema,
+)
 from ray.util.state import get_job
 from ray.dashboard.modules.job.pydantic_models import JobDetails
-from ray.util.state.common import Humanify, PredicateType
+from ray.util.state.common import Humanify
 from ray._private.gcs_utils import GcsAioClient
 import yaml
 from click.testing import CliRunner
@@ -93,9 +97,7 @@ from ray.util.state import (
     list_cluster_events,
     StateApiClient,
 )
-from ray._private.event.event_logger import get_event_id
 from ray.util.state.common import (
-    DEFAULT_LIMIT,
     DEFAULT_RPC_TIMEOUT,
     ActorState,
     ListApiOptions,
@@ -104,10 +106,8 @@ from ray.util.state.common import (
     ObjectState,
     PlacementGroupState,
     RuntimeEnvState,
-    SupportedFilterType,
     TaskState,
     WorkerState,
-    ClusterEventState,
     StateSchema,
     state_column,
 )
@@ -161,23 +161,6 @@ async def state_api_manager_e2e(ray_start_with_dashboard):
     gcs_address = address_info["gcs_address"]
     manager = get_state_api_manager(gcs_address)
     yield manager
-
-
-def verify_schema(state, result_dict: dict, detail: bool = False):
-    state_fields_columns = set()
-    if detail:
-        state_fields_columns = state.columns()
-    else:
-        state_fields_columns = state.base_columns()
-
-    for k in state_fields_columns:
-        assert k in result_dict
-
-    for k in result_dict:
-        assert k in state_fields_columns
-
-    # Make the field values can be converted without error as well
-    state(**result_dict)
 
 
 def generate_actor_data(id, state=ActorTableData.ActorState.ALIVE, class_name="class"):
@@ -346,25 +329,6 @@ def generate_runtime_env_info(runtime_env, creation_time=None, success=True):
             )
         ],
         total=1,
-    )
-
-
-def create_api_options(
-    timeout: int = DEFAULT_RPC_TIMEOUT,
-    limit: int = DEFAULT_LIMIT,
-    filters: List[Tuple[str, PredicateType, SupportedFilterType]] = None,
-    detail: bool = False,
-    exclude_driver: bool = True,
-):
-    if not filters:
-        filters = []
-    return ListApiOptions(
-        limit=limit,
-        timeout=timeout,
-        filters=filters,
-        server_timeout_multiplier=1.0,
-        detail=detail,
-        exclude_driver=exclude_driver,
     )
 
 
@@ -784,65 +748,6 @@ async def test_api_manager_list_pgs(state_api_manager):
             option=create_api_options(limit=1)
         )
     assert exc_info.value.args[0] == GCS_QUERY_FAILURE_WARNING
-
-
-@pytest.mark.asyncio
-async def test_api_manager_list_cluster_events(state_api_manager):
-    data_source_client = state_api_manager.data_source_client
-    event_id_1 = get_event_id()
-    event_id_2 = get_event_id()
-    data_source_client.get_all_cluster_events.return_value = {
-        "job_1": {
-            event_id_1: {
-                "timestamp": 10,
-                "severity": "DEBUG",
-                "message": "a",
-                "event_id": event_id_1,
-                "source_type": "GCS",
-            },
-            event_id_2: {
-                "timestamp": 10,
-                "severity": "INFO",
-                "message": "b",
-                "event_id": event_id_2,
-                "source_type": "GCS",
-            },
-        }
-    }
-    result = await state_api_manager.list_cluster_events(option=create_api_options())
-    data = result.result
-    data = data[0]
-    verify_schema(ClusterEventState, data)
-    assert result.total == 2
-
-    """
-    Test detail
-    """
-    # TODO(sang)
-
-    """
-    Test limit
-    """
-    assert len(result.result) == 2
-    result = await state_api_manager.list_cluster_events(
-        option=create_api_options(limit=1)
-    )
-    data = result.result
-    assert len(data) == 1
-    assert result.total == 2
-
-    """
-    Test filters
-    """
-    # If the column is not supported for filtering, it should raise an exception.
-    with pytest.raises(ValueError):
-        result = await state_api_manager.list_cluster_events(
-            option=create_api_options(filters=[("time", "=", "20")])
-        )
-    result = await state_api_manager.list_cluster_events(
-        option=create_api_options(filters=[("severity", "=", "INFO")])
-    )
-    assert len(result.result) == 1
 
 
 @pytest.mark.asyncio
