+++ python/ray/util/state/state_manager.py
@@ -47,11 +47,9 @@ from ray.core.generated.runtime_env_agent_pb2 import (
     GetRuntimeEnvsInfoReply,
     GetRuntimeEnvsInfoRequest,
 )
-from ray.dashboard.datacenter import DataSource
 from ray.dashboard.modules.job.common import JobInfoStorageClient
 from ray.dashboard.modules.job.pydantic_models import JobDetails, JobType
 from ray.dashboard.modules.job.utils import get_driver_jobs
-from ray.dashboard.utils import Dict as Dictionary
 from ray.util.state.common import (
     RAY_MAX_LIMIT_FROM_DATA_SOURCE,
     PredicateType,
@@ -447,9 +445,6 @@ class StateDataSourceClient:
         ]
 
         return list(driver_jobs.values()) + submission_jobs
-
-    async def get_all_cluster_events(self) -> Dictionary:
-        return DataSource.events
 
     @handle_grpc_network_errors
     async def get_task_info(
